Small changes to UserFilteringAdapter
Changed the commands to now make sense
Split up the functionality into more methods
